leprechaun/requirements.txt
# common
psycopg2==2.9.3
redis==4.3.4

# django
django==4.1
django-model-utils==4.2.0
django-jazzmin==2.5.0
django-health-check==3.17.0


# asgi


# tests

